+++ app/presenters/outgoing_transfer_presenter.rb
@@ -1,4 +1,4 @@
-class TransferPresenter < SimpleDelegator
+class OutgoingTransferPresenter < SimpleDelegator
   def value
     return if super.blank?
     ActionController::Base.helpers.number_to_currency(super, unit: "£")