+++ spec/features/staff/users_can_edit_an_outgoing_transfer_spec.rb
@@ -6,28 +6,28 @@ RSpec.feature "Delivery partner users can edit a transfer" do
   let(:destination_activity) { create(:project_activity) }
   let(:report) { Report.for_activity(source_activity).create(state: "active") }
 
-  let!(:transfer) { create(:transfer, report: report, source: source_activity, destination: destination_activity) }
+  let!(:transfer) { create(:outgoing_transfer, report: report, source: source_activity, destination: destination_activity) }
 
   before do
     visit organisation_activity_path(source_activity.organisation, source_activity)
-    find("a[href='#{edit_activity_transfer_path(source_activity.id, transfer.id)}']").click
+    find("a[href='#{edit_activity_outgoing_transfer_path(source_activity.id, transfer.id)}']").click
   end
 
   scenario "it shows the transfer details" do
-    expect(page).to have_content(t("page_title.transfer.edit"))
+    expect(page).to have_content(t("page_title.outgoing_transfer.edit"))
 
-    expect(page).to have_field("transfer[destination]", with: transfer.destination.roda_identifier)
-    expect(page).to have_field("transfer[financial_quarter]", with: transfer.financial_quarter, checked: true)
+    expect(page).to have_field("outgoing_transfer[destination]", with: transfer.destination.roda_identifier)
+    expect(page).to have_field("outgoing_transfer[financial_quarter]", with: transfer.financial_quarter, checked: true)
     expect(page).to have_selector("option[value='#{transfer.financial_year}'][selected='selected']")
-    expect(page).to have_field("transfer[value]", with: transfer.value)
+    expect(page).to have_field("outgoing_transfer[value]", with: transfer.value)
   end
 
   scenario "the user can edit a transfer succesfully" do
-    new_transfer = fill_in_transfer_form(value: 1234)
+    new_transfer = fill_in_outgoing_transfer_form(value: 1234)
 
     click_on t("default.button.submit")
 
-    expect(page).to have_content(t("page_title.transfer.confirm"))
+    expect(page).to have_content(t("page_title.outgoing_transfer.confirm"))
     expect(page).to have_content(new_transfer.destination.title)
     expect(page).to have_content(new_transfer.destination.organisation.name)
     expect(page).to have_content(FinancialQuarter.new(new_transfer.financial_year, new_transfer.financial_quarter).to_s)
@@ -35,7 +35,7 @@ RSpec.feature "Delivery partner users can edit a transfer" do
 
     click_on "Yes"
 
-    expect(page).to have_content(t("action.transfer.update.success"))
+    expect(page).to have_content(t("action.outgoing_transfer.update.success"))
 
     expect(transfer.reload.destination).to eq(new_transfer.destination)
     expect(transfer.financial_year).to eq(new_transfer.financial_year)
@@ -46,18 +46,18 @@ RSpec.feature "Delivery partner users can edit a transfer" do
   scenario "the user can edit their response" do
     click_on t("default.button.submit")
 
-    expect(page).to have_content(t("page_title.transfer.confirm"))
+    expect(page).to have_content(t("page_title.outgoing_transfer.confirm"))
 
     click_on "No"
 
-    expect(page).to have_content(t("page_title.transfer.edit"))
+    expect(page).to have_content(t("page_title.outgoing_transfer.edit"))
 
-    new_transfer = fill_in_transfer_form
+    new_transfer = fill_in_outgoing_transfer_form
     click_on t("default.button.submit")
 
     click_on "Yes"
 
-    expect(page).to have_content(t("action.transfer.update.success"))
+    expect(page).to have_content(t("action.outgoing_transfer.update.success"))
 
     expect(transfer.reload.destination).to eq(new_transfer.destination)
     expect(transfer.financial_year).to eq(new_transfer.financial_year)
@@ -71,13 +71,13 @@ RSpec.feature "Delivery partner users can edit a transfer" do
     roda_identifier = "GCRF-BLOB-424434434"
     allow(non_existent_activity).to receive(:roda_identifier) { roda_identifier }
 
-    fill_in_transfer_form(destination: non_existent_activity, value: nil)
+    fill_in_outgoing_transfer_form(destination: non_existent_activity, value: nil)
 
     click_on t("default.button.submit")
 
-    expect(page).to have_content(t("activerecord.errors.models.transfer.attributes.destination.required"))
-    expect(page).to have_content(t("activerecord.errors.models.transfer.attributes.value.blank"))
+    expect(page).to have_content(t("activerecord.errors.models.outgoing_transfer.attributes.destination.required"))
+    expect(page).to have_content(t("activerecord.errors.models.outgoing_transfer.attributes.value.blank"))
 
-    expect(page).to have_field("transfer[destination]", with: roda_identifier)
+    expect(page).to have_field("outgoing_transfer[destination]", with: roda_identifier)
   end
 end